+++ lib/presentation/widgets/common/notification_overlay_manager.dart
@@ -13,6 +13,8 @@ enum NotificationType {
   streakBroken,
   badgeEarned,
   assignment,
+  questComplete,
+  allQuestsComplete,
 }
 
 // ---------------------------------------------------------------------------